DVH #14 Small file structure refactoring
