+++ firmware/dvh/select/lab_registry.c
