+++ firmware/dvh/select/lab_registry.h
