+++ firmware/dvh/shared/src/utils_secrets.c
