+++ firmware/dvh/shared/src/utils_uart.c
