+++ firmware/tests/unit/src/test_lab_select.c
@@ -1,5 +1,5 @@
 #include "../lib/unity.h"
-#include "../../../select/lab_select.c"
+#include "../../../dvh/select/lab_select.c"
 #include <stdint.h>
 
 void setUp(void) {}
