+++ firmware/tests/unit/src/test_utils_secrets.c
@@ -1,5 +1,5 @@
 #include "../lib/unity.h"
-#include "../../../shared/src/utils_secrets.c"
+#include "../../../dvh/shared/src/utils_secrets.c"
 #include <string.h>
 #include <stdbool.h>
