+++ firmware/tests/unit/src/test_utils_shell.c
@@ -1,5 +1,5 @@
 #include "../lib/unity.h"
-#include "../../../shared/src/utils_shell.c"
+#include "../../../dvh/shared/src/utils_shell.c"
 #include "stm32f1xx_hal.h"
 #include <string.h>
