+++ firmware/tests/unit/src/test_utils_uart.c
@@ -1,5 +1,5 @@
 #include "../lib/unity.h"
-#include "../../../shared/src/utils_uart.c"
+#include "../../../dvh/shared/src/utils_uart.c"
 #include "stm32f1xx_hal.h"
 #include <string.h>
